# docs/swagger.yaml
definitions:
  main.ChirpRequest:
    properties:
      body:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: Chirpy API
paths:
...
  /api/chirps:
    get:
      responses: {}
    post:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: chirp
        in: body
        name: chirp
        required: true
        schema:
          $ref: '#/definitions/main.ChirpRequest'
      responses: {}
  /api/chirps/{chirpID}:
    delete:
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: chirp id
        in: path
        name: chirpID
        required: true
        type: string
      responses: {}
    get:
      parameters:
      - description: chirp id
        in: path
        name: chirpID
        required: true
        type: string
      responses: {}
...
swagger: "2.0"